requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
PyYAML>=6.0
tqdm>=4.66
matplotlib>=3.8
seaborn>=0.13
torch>=2.2
modal>=1.2
